Add support for terrain holes
Jolt's heightmap shape supports a special height value FLT_MAX, which
signifies the cNoCollisionValue gets handled as a hole in the terrain.

The FLT_MAX is not exposed to Lua. Instead, the heightfield callback
can return a `nil` value for x,z coordinates to specify that the
heightmap area should be ignored for collisions.