Model doesn't retain ModelData;
This allows CPU memory for model assets to be freed after they are
uploaded to VRAM, drastically reducing memory usage for projects that
load many/big models.

ModelData was split into ModelMetadata (lightweight metadata info) and
ModelData (heavyweight images and mesh data).  When creating a Model,
the model copies the metadata out from the ModelData and uploads the
images and meshes to the GPU, but no longer retains the ModelData.

Also:

- Removed Model:getData, for obvious reasons.
- Removed bounding sphere.  I don't think this is good design.  There
  are different ways to compute a bounding sphere, and you can also
  trade off precision vs. speed.  I don't think LOVR can effectively
  decide which approach to use.
- Model methods that previously called into ModelData use a more
  efficient approach, saving the 2-3us of overhead per call.